kpath: workaround for absolute paths in texmf
Paths that are built from an ls-R database are, ultimately, absolute.
But the associated kpath.Context may have been created from an fs.FS root
that won't be able to recoup the extra leading path.

Workaround this be first trying fs.FS.Open and then os.Open.

This should be fixed by a mount-FS that knows its full-path in the
overall fs.FS hierarchy.

Signed-off-by: Sebastien Binet <s@sbinet.org>